Files modified: 1
Modified 1 file:GitSyncMacApp/lib/graph2/GraphView2.swift

+++ GitSyncMacApp/lib/graph2/GraphView2.swift
@@ -31,7 +31,7 @@ extension GraphView2{
      */
     func addGraphLine(){
         addGraphLineStyle()
-        let h:CGFloat = height.int*2
+        let h:Int = height.int*2
         points = (0..<18).map{
             let x:CGFloat = 100*$0
             let y:CGFloat = (0..<(h-32)).random.cgFloat
